+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 0.1.1 (2021-12-06)
 ------------------
 
+* Implement escape sequences for save/restore cursor
 * Fix: Use correct colour palette as per http://answers.google.com/answers/threadview/id/126097.html
 * Fix: Bright flag should not apply to background colour